Give contacts its own presenter
