+++ app/src/main/java/chat/rocket/android/chatrooms/ui/ChatRoomsFragment.kt
@@ -28,7 +28,7 @@ import chat.rocket.android.chatrooms.viewmodel.ChatRoomsViewModel
 import chat.rocket.android.chatrooms.viewmodel.ChatRoomsViewModelFactory
 import chat.rocket.android.chatrooms.viewmodel.LoadingState
 import chat.rocket.android.chatrooms.viewmodel.Query
-import chat.rocket.android.contacts.ContactsFragment
+import chat.rocket.android.contacts.ui.ContactsFragment
 import chat.rocket.android.helper.ChatRoomsSortOrder
 import chat.rocket.android.helper.Constants
 import chat.rocket.android.helper.SharedPreferenceHelper
