+++ app/src/main/java/chat/rocket/android/contacts/adapter/ContactsHeaderItemHolder.kt
@@ -2,4 +2,4 @@ package chat.rocket.android.contacts.adapter
 
 import chat.rocket.android.chatrooms.adapter.ItemHolder
 
-data class inviteItemHolder(override val data: String) : ItemHolder<String>
+data class ContactsHeaderItemHolder(override val data: String) : ItemHolder<String>
