+++ app/src/main/java/chat/rocket/android/contacts/adapter/ContactsItemHolder.kt
@@ -3,4 +3,4 @@ package chat.rocket.android.contacts.adapter
 import chat.rocket.android.chatrooms.adapter.ItemHolder
 import chat.rocket.android.contacts.models.Contact
 
-data class ContactItemHolder(override val data: Contact) : ItemHolder<Contact>
+data class ContactsItemHolder(override val data: Contact) : ItemHolder<Contact>
