+++ app/src/main/java/chat/rocket/android/contacts/adapter/ContactsRecyclerViewAdapter.kt
@@ -9,16 +9,17 @@ import androidx.recyclerview.widget.RecyclerView
 import chat.rocket.android.R
 import chat.rocket.android.chatrooms.adapter.*
 import chat.rocket.android.contacts.models.Contact
+import chat.rocket.android.contacts.presentation.ContactsPresenter
 import chat.rocket.android.main.ui.MainActivity
 import chat.rocket.android.util.extensions.inflate
 import chat.rocket.common.model.UserPresence
 import kotlinx.coroutines.experimental.android.UI
 import kotlinx.coroutines.experimental.launch
 
-class ContactRecyclerViewAdapter(
+class ContactsRecyclerViewAdapter(
         private val context: MainActivity,
+        private val presenter: ContactsPresenter,
         private val contactArrayList: List<ItemHolder<*>>
-
 ) : RecyclerView.Adapter<ViewHolder<*>>() {
 
     init {
@@ -30,11 +31,11 @@ class ContactRecyclerViewAdapter(
             VIEW_TYPE_CONTACT -> {
                 val layoutInflater = LayoutInflater.from(parent.context)
                 val binding: ViewDataBinding = DataBindingUtil.inflate(layoutInflater, R.layout.item_contact, parent, false)
-                ContactViewHolder(binding.root)
+                ContactsViewHolder(binding.root)
             }
             VIEW_TYPE_HEADER -> {
                 val view = parent.inflate(R.layout.item_heading)
-                ContactHeaderViewHolder(view)
+                ContactsHeaderViewHolder(view)
             }
             VIEW_TYPE_INVITE -> {
                 val view = parent.inflate(R.layout.item_invite)
@@ -49,31 +50,31 @@ class ContactRecyclerViewAdapter(
     override fun getItemId(position: Int): Long {
         val item = contactArrayList[position]
         return when (item) {
-            is ContactItemHolder -> item.data.hashCode().toLong()
-            is ContactHeaderItemHolder -> item.data.hashCode().toLong()
-            is inviteItemHolder -> item.data.hashCode().toLong()
+            is ContactsItemHolder -> item.data.hashCode().toLong()
+            is ContactsHeaderItemHolder -> item.data.hashCode().toLong()
+            is InviteItemHolder -> item.data.hashCode().toLong()
             else -> throw IllegalStateException("View type must be either Room, Header or Invite")
         }
     }
 
     override fun getItemViewType(position: Int): Int {
         return when (contactArrayList[position]) {
-            is ContactItemHolder -> VIEW_TYPE_CONTACT
-            is ContactHeaderItemHolder -> VIEW_TYPE_HEADER
-            is inviteItemHolder -> VIEW_TYPE_INVITE
+            is ContactsItemHolder -> VIEW_TYPE_CONTACT
+            is ContactsHeaderItemHolder -> VIEW_TYPE_HEADER
+            is InviteItemHolder -> VIEW_TYPE_INVITE
             else -> throw IllegalStateException("View type must be either Room, Header or Invite")
         }
     }
 
     override fun onBindViewHolder(holder: ViewHolder<*>, position: Int) {
-        if (holder is ContactViewHolder) {
-            holder.bind(contactArrayList[position] as ContactItemHolder)
+        if (holder is ContactsViewHolder) {
+            holder.bind(contactArrayList[position] as ContactsItemHolder)
 
             val contact: Contact = holder.data!!.data
             val userId = contact.getUserId()
             if (userId != null) {
                 launch {
-                    var userPresence: UserPresence? = context.presenter.getUserPresence(userId)
+                    var userPresence: UserPresence? = presenter.getUserPresence(userId)
                     if (userPresence != null) {
                         contact.setStatus(userPresence.presence!!)
                     }
@@ -88,26 +89,26 @@ class ContactRecyclerViewAdapter(
 
             inviteButton.setOnClickListener { view ->
                 run {
-                    inviteButton.setText(context.getString(R.string.Invited))
+                    inviteButton.setText(R.string.Invited)
                     if (contact.isPhone()) {
-                        context.presenter.inviteViaSMS(contact.getPhoneNumber()!!)
+                        presenter.inviteViaSMS(contact.getPhoneNumber()!!)
                     } else {
-                        context.presenter.inviteViaEmail(contact.getEmailAddress()!!)
+                        presenter.inviteViaEmail(contact.getEmailAddress()!!)
                     }
                 }
             }
 
             dmButton.setOnClickListener { view ->
                 run {
-                    context.presenter.openDirectMessageChatRoom(contact.getUsername().toString())
+                    presenter.openDirectMessageChatRoom(contact.getUsername().toString())
                 }
             }
 
-         } else if (holder is ContactHeaderViewHolder) {
-            holder.bind(contactArrayList[position] as ContactHeaderItemHolder)
+         } else if (holder is ContactsHeaderViewHolder) {
+            holder.bind(contactArrayList[position] as ContactsHeaderItemHolder)
 
         } else if (holder is InviteViewHolder) {
-            holder.bind(contactArrayList[position] as inviteItemHolder)
+            holder.bind(contactArrayList[position] as InviteItemHolder)
             holder.itemView.setOnClickListener {
                 shareApp()
             }
@@ -115,7 +116,7 @@ class ContactRecyclerViewAdapter(
     }
 
     private fun shareApp() {
-        context.presenter.shareViaApp(context)
+        presenter.shareViaApp(context)
     }
 
     companion object {
