+++ app/src/main/java/chat/rocket/android/contacts/adapter/InviteItemHolder.kt
@@ -2,4 +2,4 @@ package chat.rocket.android.contacts.adapter
 
 import chat.rocket.android.chatrooms.adapter.ItemHolder
 
-data class ContactHeaderItemHolder(override val data: String) : ItemHolder<String>
+data class InviteItemHolder(override val data: String) : ItemHolder<String>
