+++ app/src/main/java/chat/rocket/android/contacts/adapter/InviteViewHolder.kt
@@ -3,7 +3,7 @@ package chat.rocket.android.contacts.adapter
 import android.view.View
 import chat.rocket.android.chatrooms.adapter.ViewHolder
 
-class InviteViewHolder(itemView: View) : ViewHolder<inviteItemHolder>(itemView) {
-    override fun bindViews(data: inviteItemHolder) {
+class InviteViewHolder(itemView: View) : ViewHolder<InviteItemHolder>(itemView) {
+    override fun bindViews(data: InviteItemHolder) {
     }
 }
