+++ app/src/main/java/chat/rocket/android/contacts/di/ContactsFragmentProvider.kt
@@ -1,0 +1,14 @@
+package chat.rocket.android.contacts.di
+
+import chat.rocket.android.contacts.ui.ContactsFragment
+import chat.rocket.android.dagger.scope.PerFragment
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+
+@Module
+abstract class ContactsFragmentProvider {
+
+    @ContributesAndroidInjector(modules = [ContactsFragmentModule::class])
+    @PerFragment
+    abstract fun provideContactsFragment(): ContactsFragment
+}
