+++ app/src/main/java/chat/rocket/android/contacts/presentation/ContactsView.kt
@@ -1,0 +1,7 @@
+package chat.rocket.android.contacts.presentation
+
+import chat.rocket.android.core.behaviours.LoadingView
+import chat.rocket.android.core.behaviours.MessageView
+
+interface ContactsView : LoadingView, MessageView {
+}
