+++ app/src/main/java/chat/rocket/android/dagger/module/WorkerBuilder.kt
@@ -1,8 +1,8 @@
 package chat.rocket.android.dagger.module
 
 import androidx.work.Worker
-import chat.rocket.android.contacts.di.ContactSyncSubComponent
-import chat.rocket.android.contacts.worker.ContactSyncWorker
+import chat.rocket.android.contacts.di.ContactsSyncSubComponent
+import chat.rocket.android.contacts.worker.ContactsSyncWorker
 import chat.rocket.android.dagger.qualifier.WorkerKey
 import dagger.Binds
 import dagger.Module
@@ -10,12 +10,12 @@ import dagger.android.AndroidInjector
 import dagger.multibindings.IntoMap
 
 
-@Module(subcomponents = [ContactSyncSubComponent::class])
+@Module(subcomponents = [ContactsSyncSubComponent::class])
 abstract class WorkerBuilder {
     @Binds
     @IntoMap
-    @WorkerKey(ContactSyncWorker::class)
-    abstract fun bindContactSyncWorkerFactory(
-            builder: ContactSyncSubComponent.Builder
+    @WorkerKey(ContactsSyncWorker::class)
+    abstract fun bindContactsSyncWorkerFactory(
+            builder: ContactsSyncSubComponent.Builder
     ): AndroidInjector.Factory<out Worker>
 }
